how many open files

+++ ch03/ppid.c
@@ -1,5 +1,6 @@
 // ppid is a simple example to see a system call in assembly code
 #include <stdio.h>
+#include <stdint.h>
 #include <sys/types.h>
 #include <unistd.h>
 
@@ -7,5 +8,5 @@ int main() {
 	pid_t ppid;
 
 	ppid = getppid();
-	printf("%d\n", ppid);
+	printf("%jd\n", (intmax_t)ppid);
 }
